[refactor](be) Organize new parquet column readers
### What problem does this PR solve?

Issue Number: close #xxx

Related PR: #xxx

Problem Summary: Move new parquet column reader implementations under a dedicated column_reader directory and split struct/list/map reader declarations into independent headers without keeping the complex reader umbrella header.

### Release note

None

### Check List (For Author)

- Test: Manual test

    - build-support/clang-format.sh

    - git diff --check

    - BUILD_TYPE=DEBUG ./build.sh --be on Fedora

- Behavior changed: No

- Does this need documentation: Yes (updated docs/doris-new-parquet-reader-internal-layering.md)
