+++ be/src/format/new_parquet/column_reader/scalar_column_reader.h
@@ -20,8 +20,8 @@
 #include <memory>
 #include <string>
 
-#include "format/new_parquet/arrow_leaf_reader_adapter.h"
-#include "format/new_parquet/column_reader.h"
+#include "format/new_parquet/column_reader/arrow_leaf_reader_adapter.h"
+#include "format/new_parquet/column_reader/column_reader.h"
 #include "format/new_parquet/parquet_type.h"
 
 namespace parquet {
