+++ be/src/format/new_parquet/column_reader/struct_column_reader.cpp
@@ -15,15 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
+#include "format/new_parquet/column_reader/struct_column_reader.h"
+
 #include <algorithm>
 #include <cstdint>
 #include <utility>
 #include <vector>
 
 #include "core/column/column_struct.h"
-#include "format/new_parquet/complex_column_reader.h"
-#include "format/new_parquet/complex_column_reader_helpers.h"
-#include "format/new_parquet/scalar_column_reader.h"
+#include "format/new_parquet/column_reader/complex_column_reader_helpers.h"
+#include "format/new_parquet/column_reader/scalar_column_reader.h"
 
 namespace doris::parquet {
